Create component of adding new product

+++ src/main/java/by/epam/project/controller/async/command/impl/RemoveProfileImageCommand.java
@@ -5,6 +5,8 @@ import by.epam.project.exception.ServiceException;
 import by.epam.project.model.entity.User;
 import by.epam.project.model.service.UserService;
 import by.epam.project.model.service.impl.UserServiceImpl;
+import by.epam.project.util.FileUtil;
+import by.epam.project.util.JsonUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -12,10 +14,15 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import java.io.File;
+import java.io.IOException;
+import java.util.Optional;
+
+import static by.epam.project.controller.parameter.ParameterKey.URL;
 import static by.epam.project.controller.parameter.ParameterKey.USER;
 
 
-public class DeleteProfileImage implements Command {
+public class RemoveProfileImageCommand implements Command {
     private static final Logger logger = LogManager.getLogger();
 
     private final UserService userService = UserServiceImpl.getInstance();
@@ -28,8 +35,17 @@ public class DeleteProfileImage implements Command {
         String login = user.getLogin();
 
         try {
+            Optional<String> URLOptional = userService.findAvatarURLByLogin(login);
+
+            if (URLOptional.isEmpty()) {
+                response.setStatus(HttpServletResponse.SC_NOT_FOUND);
+                return;
+            }
+
+            String avatarURL = URLOptional.get();
+            FileUtil.remove(avatarURL);
             userService.removeAvatarByLogin(login);
-        } catch (ServiceException exp) {
+        } catch (ServiceException | IOException exp) {
             logger.error(exp);
             response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
         }
